separate comparisons and tracking benchmarks
